Complete tutorial for CPR, allow return to menu
A user path has been made from the breathing part of the tutorial all
the way to the completion of 30 compressions and a dialogue box allowing
return to the main menu. The user is able to repeat the tutorial over
and over again, since the last button of the tutorial resets every
object to its original tutorial state.